fix contract verification

+++ components/txtx-addon-network-evm/src/tests/fixtures/foundry/src/SimpleStorage.sol
@@ -6,24 +6,24 @@ contract SimpleStorage {
 
   struct People {
     string name;
-    uint256 favoriteNumber;
+    uint16 favoriteNumber;
   }
 
-  mapping(string => uint256) public nameToFavoriteNumber;
+  mapping(string => uint16) public nameToFavoriteNumber;
 
-  uint256 favoriteNumber;
+  uint16 favoriteNumber;
 
   People[] public people;
 
-  function store(uint256 _favoriteNumber) public virtual {
+  function store(uint16 _favoriteNumber) public virtual {
     favoriteNumber = _favoriteNumber;
   }
 
-  function retrieve() public view returns (uint256) {
+  function retrieve() public view returns (uint16) {
     return favoriteNumber;
   }
 
-  function addPerson(string memory _name, uint256 _favoriteNumber) public {
+  function addPerson(string memory _name, uint16 _favoriteNumber) public {
     people.push(People({favoriteNumber: _favoriteNumber, name: _name}));
     nameToFavoriteNumber[_name] = _favoriteNumber;
   }
